requirements.txt
numpy>=1.23
pandas>=2.0
pyyaml>=6.0
openpyxl>=3.1
scikit-learn>=1.3

fastapi
uvicorn
requests
flask

langchain-openai
langchain-anthropic
langchain-google-genai
langchain-core
langgraph

# earthshaker  # Package not found - commented out